Fixed unexpected scale of output
When you have a comp layer with 110% scale and apply 2dCam output to it, the scale would not match the 2dCam. So now, the comp layer's scale will not be used.

Alternatively, we could just set the scale property to 100% and delete the keyframes of that property, but the other properties aren't taken into account anyway